Tweak expectations for CrossObjectAccessTest
We should keep executions of rules that didn't pass the guard.
In the regular EVL, the ConstraintTrace only records the executions
of rules for which the guard returned true. In the EVL trace model,
we want record executions that did not pass the guard, as we will
use their property accesses to decide whether a change invalidated
the guard's result, therefore requiring a re-evaluation of the
guard.